+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Change Log
 
+## 0.8.0
+
+- Add support for Drag and Drop behind the `dnd` feature
+- **Breaking** Introduced generic surface type parameter for `Clipboard`.
+- Add support for loading and storing custom data.
+- **Breaking** `Clipboard::load` renamed to `Clipboard::load_text`.
+- **Breaking** `Clipboard::load_primary` renamed to `Clipboard::load_primary_text`.
+- **Breaking** `Clipboard::store` renamed to `Clipboard::store_text`.
+- **Breaking** `Clipboard::store_primary` renamed to `Clipboard::store_primary_text`.
+
 ## 0.7.1
 
 - Don't panic on display disconnect